backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
python-multipart>=0.0.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
google-cloud-documentai==2.20.1
google-cloud-storage==2.14.0
pandas==2.1.3
openpyxl==3.1.2
rapidfuzz==3.5.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
python-dateutil==2.8.2

